fix: collection drop down
