+++ components/CreateSuccess/CreatedMomentAirdrop.tsx
@@ -4,13 +4,13 @@ import MomentAirdrop from "../MomentAirdrop";
 import { Address } from "viem";
 import { MomentProvider } from "@/providers/MomentProvider";
 import { CHAIN_ID } from "@/lib/consts";
-import useCollectionParam from "@/hooks/useCollectionParam";
+import { useCollectionsProvider } from "@/providers/CollectionsProvider";
 import { Fragment } from "react";
 import useTokenIdParam from "@/hooks/useTokenIdParam";
 
 const CreatedMomentAirdrop = () => {
   const tokenId = useTokenIdParam();
-  const collection = useCollectionParam();
+  const { selectedCollection: collection } = useCollectionsProvider();
 
   if (!collection) return <Fragment />;
