--- hooks/useCollectionParam.ts
@@ -1,9 +1,0 @@
-import { useSearchParams } from "next/navigation";
-
-const useCollectionParam = () => {
-  const searchParams = useSearchParams();
-  const collection = searchParams.get("collectionAddress") as string;
-  return collection;
-};
-
-export default useCollectionParam;
